rtp-recv: Allow same timestamp in consecutive packets
The current code is built with the assumption that the timestamp field
in one RTP header will be greater than the one from the preceding packet,
which is true when recieving from another PA instance, but not in the
general case.

According to RFC 3550 section 5.1: Several consecutive RTP packets will
have equal timestamps if they are (logically) generated at once, e.g.,
belong to the same video frame.

One example of a software producing this kind of stream is VLC.